moving types_*.h headers into types/ include directory
