+++ channels/cliprdr/cliprdr_main.c
@@ -23,7 +23,6 @@
 #include <string.h>
 #include <pthread.h>
 #include <unistd.h>
-#include <freerdp/types_ui.h>
 #include <freerdp/vchan.h>
 #include <freerdp/utils/stream.h>
 #include <freerdp/utils/chan_plugin.h>
