+++ channels/drdynvc/tsmf/ffmpeg/tsmf_ffmpeg.c
@@ -21,7 +21,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <libavcodec/avcodec.h>
-#include <freerdp/constants_ui.h>
+#include <freerdp/constants/ui.h>
 #include "tsmf_constants.h"
 #include "tsmf_decoder.h"
