+++ channels/rdpsnd/alsa/rdpsnd_alsa.c
@@ -22,7 +22,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <alsa/asoundlib.h>
-#include <freerdp/types_ui.h>
+#include <freerdp/types/ui.h>
 #include "rdpsnd_types.h"
 #include <freerdp/utils/stream.h>
 #include <freerdp/utils/wait_obj.h>
