+++ channels/rdpsnd/pulse/rdpsnd_pulse.c
@@ -20,7 +20,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <pulse/pulseaudio.h>
-#include <freerdp/types_ui.h>
+#include <freerdp/types/ui.h>
 #include "rdpsnd_types.h"
 #include <freerdp/utils/stream.h>
 #include <freerdp/utils/wait_obj.h>
