+++ include/freerdp/types/base.h
